requirements.txt
asgiref==3.5.0
Django==4.0.4
django-etc==1.3.1
django-hitcount==1.3.5
django-utils-six==2.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
